--- scripts/random_baseline.py
@@ -1,15 +1,0 @@
-import random, time
-from typing import Union, Optional
-from random import choice, randint
-from string import ascii_lowercase
-
-
-def get_random_string(
-    length: Optional[int] = None,
-    variability: Optional[int] = None,
-) -> str:
-    if length is None:
-        length = randint(3, 300)
-    if variability is None:
-        variability = len(ascii_lowercase)
-    return "".join(choice(ascii_lowercase[:variability]) for _ in range(length))
