+++ scripts/test.js
