+++ scripts/test.py
@@ -1,14 +1,115 @@
-import random, time
-from typing import Union, Optional
-from random import choice, randint
-from string import ascii_lowercase
-
-import numpy as np
 import pytest
 
 import stringzilla as sz
-from stringzilla import Str, Strs
-from scripts.similarity_baseline import levenshtein
+from stringzilla import Str
+
+
+def test_unit_construct():
+    native = "aaaaa"
+    big = Str(native)
+    assert len(big) == len(native)
+
+
+def test_unit_indexing():
+    native = "abcdef"
+    big = Str(native)
+    for i in range(len(native)):
+        assert big[i] == native[i]
+
+
+def test_unit_count():
+    native = "aaaaa"
+    big = Str(native)
+    assert big.count("a") == 5
+    assert big.count("aa") == 2
+    assert big.count("aa", allowoverlap=True) == 4
+
+
+def test_unit_contains():
+    big = Str("abcdef")
+    assert "a" in big
+    assert "ab" in big
+    assert "xxx" not in big
+
+
+def test_unit_rich_comparisons():
+    assert Str("aa") == "aa"
+    assert Str("aa") < "b"
+    assert Str("abb")[1:] == "bb"
+
+
+def test_unit_buffer_protocol():
+    import numpy as np
+
+    my_str = Str("hello")
+    arr = np.array(my_str)
+    assert arr.dtype == np.dtype("c")
+    assert arr.shape == (len("hello"),)
+    assert "".join([c.decode("utf-8") for c in arr.tolist()]) == "hello"
+
+
+def test_unit_split():
+    native = "token1\ntoken2\ntoken3"
+    big = Str(native)
+    assert native.splitlines() == list(big.splitlines())
+    assert native.splitlines(True) == list(big.splitlines(keeplinebreaks=True))
+    assert native.split("token3") == list(big.split("token3"))
+
+    words = sz.split(big, "\n")
+    assert len(words) == 3
+    assert str(words[0]) == "token1"
+    assert str(words[2]) == "token3"
+
+    parts = sz.split(big, "\n", keepseparator=True)
+    assert len(parts) == 3
+    assert str(parts[0]) == "token1\n"
+    assert str(parts[2]) == "token3"
+
+
+def test_unit_sequence():
+    native = "p3\np2\np1"
+    big = Str(native)
+
+    lines = big.splitlines()
+    assert [2, 1, 0] == list(lines.order())
+
+    lines.sort()
+    assert [0, 1, 2] == list(lines.order())
+    assert ["p1", "p2", "p3"] == list(lines)
+
+    # Reverse order
+    assert [2, 1, 0] == list(lines.order(reverse=True))
+    lines.sort(reverse=True)
+    assert ["p3", "p2", "p1"] == list(lines)
+
+
+def test_unit_globals():
+    """Validates that the previously unit-tested member methods are also visible as global functions."""
+
+    assert sz.find("abcdef", "bcdef") == 1
+    assert sz.find("abcdef", "x") == -1
+
+    assert sz.count("abcdef", "x") == 0
+    assert sz.count("aaaaa", "a") == 5
+    assert sz.count("aaaaa", "aa") == 2
+    assert sz.count("aaaaa", "aa", allowoverlap=True) == 4
+
+    assert sz.edit_distance("aaa", "aaa") == 0
+    assert sz.edit_distance("aaa", "bbb") == 3
+    assert sz.edit_distance("abababab", "aaaaaaaa") == 4
+    assert sz.edit_distance("abababab", "aaaaaaaa", 2) == 2
+    assert sz.edit_distance("abababab", "aaaaaaaa", bound=2) == 2
+
+
+def get_random_string(
+    length: Optional[int] = None,
+    variability: Optional[int] = None,
+) -> str:
+    if length is None:
+        length = randint(3, 300)
+    if variability is None:
+        variability = len(ascii_lowercase)
+    return "".join(choice(ascii_lowercase[:variability]) for _ in range(length))
 
 
 def is_equal_strings(native_strings, big_strings):
@@ -79,7 +180,7 @@ def test_fuzzy_substrings(pattern_length: int, haystack_length: int, variability
 
 @pytest.mark.repeat(100)
 @pytest.mark.parametrize("max_edit_distance", [150])
-def test_levenshtein_insertions(max_edit_distance: int):
+def test_edit_distance_insertions(max_edit_distance: int):
     # Create a new string by slicing and concatenating
     def insert_char_at(s, char_to_insert, index):
         return s[:index] + char_to_insert + s[index:]
@@ -90,14 +191,42 @@ def test_levenshtein_insertions(max_edit_distance: int):
         source_offset = randint(0, len(ascii_lowercase) - 1)
         target_offset = randint(0, len(b) - 1)
         b = insert_char_at(b, ascii_lowercase[source_offset], target_offset)
-        assert sz.levenshtein(a, b, 200) == i + 1
+        assert sz.edit_distance(a, b, 200) == i + 1
 
 
 @pytest.mark.repeat(1000)
-def test_levenshtein_randos():
+def test_edit_distance_randos():
     a = get_random_string(length=20)
     b = get_random_string(length=20)
-    assert sz.levenshtein(a, b, 200) == levenshtein(a, b)
+    assert sz.edit_distance(a, b, 200) == edit_distance(a, b)
+
+
+@pytest.mark.parametrize("list_length", [10, 20, 30, 40, 50])
+@pytest.mark.parametrize("part_length", [5, 10])
+@pytest.mark.parametrize("variability", [2, 3])
+def test_fuzzy_sorting(list_length: int, part_length: int, variability: int):
+    native_list = [
+        get_random_string(variability=variability, length=part_length)
+        for _ in range(list_length)
+    ]
+    native_joined = ".".join(native_list)
+    big_joined = Str(native_joined)
+    big_list = big_joined.split(".")
+
+    native_ordered = sorted(native_list)
+    native_order = big_list.order()
+    for i in range(list_length):
+        assert native_ordered[i] == native_list[native_order[i]], "Order is wrong"
+        assert native_ordered[i] == str(
+            big_list[int(native_order[i])]
+        ), "Split is wrong?!"
+
+    native_list.sort()
+    big_list.sort()
+
+    assert len(native_list) == len(big_list)
+    for native_str, big_str in zip(native_list, big_list):
+        assert native_str == str(big_str), "Order is wrong"
 
 
 @pytest.mark.parametrize("list_length", [10, 20, 30, 40, 50])
